style: move file to `/types`
